+++ packages/in_app_purchase/example/android/app/src/test/java/io/flutter/plugins/inapppurchase/InAppPurchasePluginTest.java
@@ -29,6 +29,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import android.app.Activity;
+import android.content.Context;
 import androidx.annotation.Nullable;
 import com.android.billingclient.api.BillingClient;
 import com.android.billingclient.api.BillingClient.BillingResponse;
@@ -58,36 +59,22 @@ import org.mockito.MockitoAnnotations;
 import org.mockito.Spy;
 
 public class InAppPurchasePluginTest {
-  InAppPurchasePlugin plugin;
+  private InAppPurchasePlugin plugin;
   @Mock BillingClient mockBillingClient;
   @Mock MethodChannel mockMethodChannel;
   @Spy Result result;
   @Mock PluginRegistry.Registrar registrar;
   @Mock Activity activity;
-  @Mock BinaryMessenger messenger;
   @Mock BillingClientFactory factory;
+
 
   @Before
   public void setUp() {
     MockitoAnnotations.initMocks(this);
     when(factory.createBillingClient(any(), any())).thenReturn(mockBillingClient);
-
-    plugin = new InAppPurchasePlugin(factory, activity, mockMethodChannel);
+    plugin = new InAppPurchasePlugin(factory, registrar, mockMethodChannel);
   }
 
-  @Test
-  public void doNotRegisterWithoutActivityAvailable() {
-    InAppPurchasePlugin.registerWith(registrar);
-    verify(registrar, never()).messenger();
-  }
-
-  @Test
-  public void registerWithoutActivityAvailable() {
-    when(registrar.messenger()).thenReturn(messenger);
-    when(registrar.activity()).thenReturn(activity);
-    InAppPurchasePlugin.registerWith(registrar);
-    verify(registrar).messenger();
-  }
 
   @Test
   public void invalidMethod() {
@@ -236,6 +223,7 @@ public class InAppPurchasePluginTest {
 
   @Test
   public void launchBillingFlow_ok_nullAccountId() {
+    when(registrar.activity()).thenReturn(activity);
     // Fetch the sku details first and then prepare the launch billing flow call
     String skuId = "foo";
     queryForSkus(singletonList(skuId));
@@ -263,7 +251,26 @@ public class InAppPurchasePluginTest {
   }
 
   @Test
+  public void launchBillingFlow_ok_null_Activity() {
+    // Fetch the sku details first and then prepare the launch billing flow call
+    String skuId = "foo";
+    String accountId = "account";
+    queryForSkus(singletonList(skuId));
+    HashMap<String, Object> arguments = new HashMap<>();
+    arguments.put("sku", skuId);
+    arguments.put("accountId", accountId);
+    MethodCall launchCall = new MethodCall(LAUNCH_BILLING_FLOW, arguments);
+
+    plugin.onMethodCall(launchCall, result);
+
+    // Verify we pass the response code to result
+    verify(result).error(contains("ACTIVITY_UNAVAILABLE"), contains("foreground"), any());
+    verify(result, never()).success(any());
+  }
+
+  @Test
   public void launchBillingFlow_ok_AccountId() {
+    when(registrar.activity()).thenReturn(activity);
     // Fetch the sku details first and query the method call
     String skuId = "foo";
     String accountId = "account";